feta: add parse_validator_info
