+++ account-decoder/src/lib.rs
@@ -13,6 +13,7 @@ pub mod parse_stake;
 pub mod parse_sysvar;
 pub mod parse_token;
 pub mod parse_token_extension;
+pub mod parse_validator_info;
 pub mod parse_vote;
 pub mod validator_info;
